requirements.txt
fastapi
uvicorn[standard]
python-multipart
psycopg2-binary
python-dotenv
celery
